+++ src/components/layout/SideNavigation.module.scss
@@ -1,0 +1,3 @@
+.menuItem {
+  margin-left: 16px;
+}
